WIP service extension handling static method
This'll do the main updates of the incoming notification content - namely, apply an iOS notification category based on the Snapyr push template ID, and re-fetch category definitions from the SDK config endpoint if something's stale or missing.
We'll also want to put image download in here so it's a more full-service thing, but saving that for later.

TODO: clean up all the debugging junk, and get in-notification-processing category updates working correctly