Split in-app content into its own class; tweaks
Makes interface for checking in-app messages more in line with
that of the Android SDK, which also enforces checks a bit more.

Consumer needs to call the appropriate content getter method
depending on whether the payload type is JSON or HTML.

Also updates a couple of names to be more consistent with other
parts of the system, namely `payloadType` rather than
`contentType`.