Fix Xcode compiler errors - extension/cocoapods
When installing pod and using Snapyr in a notification extension,
getting error "'openURL:options:completionHandler:' is
unavailable: not available on iOS (App Extension)"

This method should already never be called from within an app
extension. Adding compiler flag to mark it as such so that
compilation succeeds. Also sets project "extension safe API"
setting back to on now that this satisfies the compiler.

TODO: check if there are any edge cases to this - there is an
alternative available using selectors on this method