appDidReceiveNotificationResponse: open deep links
Attempts to automatically open deep link on tapped notification, if
one was present.

If an app (including this app) matches the URL scheme on the link, it
will launch that. Else if it's an http(s) link, it should automatically
attempt to open in Safari. Else, do nothing and log error