Push handling updates (#48)
* Improve push support - helpers; notification class

Adds static methods to SnapyrSDK that can be called as one-liners
from consumer's AppDelegate for push-related events:
- push received
- push response (tap)
- push token received

Adds a SnapyrNotification class for representation push notification
data arriving from Snapyr - this corresponds to a similar class in
the Android SDK and is compatible with that class when serialized
and passed over to React Native.

* Debug log push token

* appDidReceiveNotificationResponse: open deep links

Attempts to automatically open deep link on tapped notification, if
one was present.

If an app (including this app) matches the URL scheme on the link, it
will launch that. Else if it's an http(s) link, it should automatically
attempt to open in Safari. Else, do nothing and log error

* Cleanup; comment out unused code causing warnings

* Fix compiler errors/warnings

* Lint fixes